chore: e2e tests

+++ test/e2e/e2e_suite_test.go
@@ -34,15 +34,17 @@ import (
 var (
 	// Optional Environment Variables:
 	// - CERT_MANAGER_INSTALL_SKIP=true: Skips CertManager installation during test setup.
+	// - HORIZON_ISSUER_E2E_SKIP_CLEANUP=true: Do not remove any resources at the end of tests.
 	// These variables are useful if CertManager is already installed, avoiding
 	// re-installation and conflicts.
 	skipCertManagerInstall = os.Getenv("CERT_MANAGER_INSTALL_SKIP") == "true"
+	skipCleanup            = os.Getenv("HORIZON_ISSUER_E2E_SKIP_CLEANUP") == "true"
 	// isCertManagerAlreadyInstalled will be set true when CertManager CRDs be found on the cluster
 	isCertManagerAlreadyInstalled = false
 
 	// projectImage is the name of the image which will be build and loaded
 	// with the code source changes to be tested.
-	projectImage = "example.com/horizon-issuer:v0.0.1"
+	projectImage = "quay.io/evertrust/horizon-issuer:v0.0.1"
 )
 
 // TestE2E runs the end-to-end (e2e) test suite for the project. These tests execute in an isolated,
@@ -81,12 +83,18 @@ var _ = BeforeSuite(func() {
 			_, _ = fmt.Fprintf(GinkgoWriter, "WARNING: CertManager is already installed. Skipping installation...\n")
 		}
 	}
+
+	Expect(utils.InstallHorizon()).To(Succeed(), "Failed to install Horizon")
 })
 
 var _ = AfterSuite(func() {
 	// Teardown CertManager after the suite if not skipped and if it was not already installed
-	if !skipCertManagerInstall && !isCertManagerAlreadyInstalled {
+	if !skipCertManagerInstall && !skipCleanup && !isCertManagerAlreadyInstalled {
 		_, _ = fmt.Fprintf(GinkgoWriter, "Uninstalling CertManager...\n")
 		utils.UninstallCertManager()
+	}
+
+	if !skipCleanup {
+		utils.UninstallHorizon()
 	}
 })
